requirements.txt
aircv==1.4.6
box
certifi==2021.5.30
cffi 
cycler==0.10.0
fonttools==4.25.0
imageio==2.9.0
imgaug==0.4.0
kiwisolver 
matplotlib 
mkl-fft==1.3.0
mkl-random==1.1.0
mkl-service==2.3.0
MouseInfo==0.1.3
msgpack==1.0.2
munch==2.5.0
munkres==1.1.4
networkx==2.6.3
numpy 
olefile==0.46
opencv-python==4.5.4.58
Pillow
PyAutoGUI==0.9.53
pycparser 
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing 
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.27
python-box==5.4.1
python-dateutil 
PyTweening==1.0.3
PyWavelets==1.1.1
pywin32==301
PyYAML==5.4.1
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
scikit-image==0.18.3
scipy==1.7.1
Shapely==1.8.0
sip==4.19.13
six 
tifffile==2021.10.12
toml==0.10.2
torch==1.9.0
torchaudio==0.9.0
torchvision==0.10.0
tornado 
tqdm 
typing-extensions 
wincertstore==0.2
